p2p/test-utils: Add create_two_mock_peers() utility function
